added line charts for weighted and probability callers

+++ src/main.py
@@ -30,37 +30,41 @@ if os.path.isfile(df_file):
 else:
     df = parser.combine('2017-06-30_NgsReviewer_master', bed_file, samples_dir)
 
+combined_file = os.path.join(output_dir, 'parsed_combined.tab')
 # Get small panel coverage file
 panel = parser.parse_bed(bed_file)
-
-# DataFrame terminal display options
-pandas.set_option('display.max_columns', 7)
-pandas.set_option('display.max_rows', 13)
-pandas.set_option('display.width', 300)
-
-# Generate combined callers
-#if os.path.isfile(weights_file):
-#    weights_df = pandas.read_csv(weights_file, sep='\t')
-#    weights = weights_df['WEIGHT']
-#else:
-#    weights = analysis.generate_combined_caller_weights(df)
-#    # weights_df = pandas.DataFrame({'CALLER': callers, 'WEIGHT': list(weights)})
-#    # weights_df.to_csv(weights_file, sep='\t', encoding='utf-8', index=False)
-print('Adding combined callers...')
-analysis.add_x_or_more(df)
-analysis.add_prob_caller(df)
-analysis.add_weight_caller(df)
-print('Classifying combined calls...')
-parser.classify(df, parser.get_new_caller_names(df))
-print('Creating new tab file...')
-df.to_csv(
-        os.path.join(output_dir, 'parsed_combined.tab'), sep='\t', 
-        encoding='utf-8', index=False
-)
+    
+if os.path.isfile(combined_file):
+    df = pandas.read_csv(combined_file, sep='\t', low_memory=False)
+else:
+    # DataFrame terminal display options
+    pandas.set_option('display.max_columns', 7)
+    pandas.set_option('display.max_rows', 13)
+    pandas.set_option('display.width', 300)
+    
+    # Generate combined callers
+    #if os.path.isfile(weights_file):
+    #    weights_df = pandas.read_csv(weights_file, sep='\t')
+    #    weights = weights_df['WEIGHT']
+    #else:
+    #    weights = analysis.generate_combined_caller_weights(df)
+    #    # weights_df = pandas.DataFrame({'CALLER': callers, 'WEIGHT': list(weights)})
+    #    # weights_df.to_csv(weights_file, sep='\t', encoding='utf-8', index=False)
+    print('Adding combined callers...')
+    analysis.add_x_or_more(df)
+    analysis.add_prob_caller(df)
+    analysis.add_weight_caller(df)
+    print('Classifying combined calls...')
+    parser.classify(df, parser.get_new_caller_names(df))
+    print('Creating new tab file...')
+    df.to_csv(
+            os.path.join(output_dir, 'parsed_combined.tab'), sep='\t', 
+            encoding='utf-8', index=False
+    )
 
 # print(weights)
 analysis_df = analysis.analyze_callers(df, panel)
-analysis.plot_callers(analysis_df, output_dir)
+analysis.plot_callers(df, analysis_df, output_dir)
 analysis_file = os.path.join(output_dir, 'analysis.csv')
 try:
 #    df.to_csv(
